--- DMSServer/src/DB/ApplicationParser.java
@@ -1,5 +1,0 @@
-package DB;
-
-public class ApplicationParser {
-
-}
