--- DMSServer/src/DB/DocumentParser.java
@@ -1,5 +1,0 @@
-package DB;
-
-public class DocumentParser {
-
-}
