--- DMSServer/src/DB/RoomParser.java
@@ -1,5 +1,0 @@
-package DB;
-
-public class RoomParser {
-
-}
