--- DMSServer/src/DB/ScoreParser.java
@@ -1,5 +1,0 @@
-package DB;
-
-public class ScoreParser {
-
-}
