+++ DMSServer/src/PayCheck.java
@@ -1,5 +1,3 @@
-package logic;
-
 import java.io.*;
 import java.sql.Connection;
 import java.sql.DriverManager;
