--- DMSServer/src/batch/test.java
@@ -1,11 +1,0 @@
-package batch;
-
-import java.sql.SQLException;
-
-public class test {
-	public static void main(String[] args) throws SQLException, ClassNotFoundException
-	{
-		AssignAlgorithm.batchStart();
-	}
-}
-
